Refactor Header into HeaderButton
