+++ packages/scrabble-solver/src/components/Results/types.ts
@@ -1,0 +1,7 @@
+import { ResultColumn, TranslationKey } from 'types';
+
+export interface Column {
+  className: string;
+  id: ResultColumn;
+  translationKey: TranslationKey;
+}
